--- Inneal/Views/Chat/NewChatView.swift
@@ -1,787 +1,0 @@
-//
-//  NewChatView.swift
-//  Inneal
-//
-//  Created by Brad Root on 8/30/25.
-//
-
-import SwiftData
-import SwiftUI
-
-struct NewChatView: View {
-    private let chat: Chat
-    private var userSettings: UserSettings
-    private var viewModel: ChatView.ViewModel
-    @Environment(\.modelContext) private var modelContext
-
-    @Query private var messages: [ChatMessage]
-    @State private var newMessage: String = ""
-    @State private var showPendingMessage: Bool = false
-    @State private var isPendingAlternate: Bool = false
-    @State private var statusMessage: String = "Sending message..."
-    @State private var opacityLevel = 0.0
-
-    @State private var showingSettingsSheet: Bool = false
-    @State private var showingConfirmationDialog: Bool = false
-    @State private var showingChatlog: Bool = false
-    @State private var batchEditModeEnabled: Bool = false
-    @State private var selectedForDeletion: Set<ChatMessage> = .init()
-    @State private var selectedCharacter: Character?
-    @State private var keyboardShowing: Bool = false
-    @FocusState private var isTextFieldFocused: Bool
-
-    @State private var textToEdit: String = ""
-    @State private var showTextEditor: Bool = false
-    @State private var alternateTextToEdit: String = ""
-    @State private var showAlternateTextEditor: Bool = false
-    @State private var alternateBeingEdited: ContentAlternate?
-    @State private var messageBeingEdited: ChatMessage?
-
-    @State private var showRequestDetails: Bool = false
-    @State private var requestDetails: String = ""
-    @State private var responseDetails: String = ""
-
-    @State private var currentAlternateIndex: Int = -1
-
-    @Namespace var unionNamespace
-    @Namespace var buttonUnionNamespace
-
-    init(for chat: Chat, modelContext: ModelContext, userSettings: UserSettings)
-    {
-        Log.debug("Init ChatView for \(chat.name)")
-        self.chat = chat
-        let id = chat.uuid
-        _messages = Query(
-            filter: #Predicate { $0.chatUUID == id },
-            sort: \.dateCreated
-        )
-        self.userSettings = userSettings
-        viewModel = ChatView.ViewModel(
-            for: chat,
-            modelContext: modelContext,
-            userSettings: userSettings
-        )
-    }
-
-    var body: some View {
-        ScrollViewReader { proxy in
-            ScrollView {
-                VStack {
-                    ForEach(messages) { message in
-                        HStack(alignment: .center, spacing: 10) {
-                            if batchEditModeEnabled {
-                                Button {
-                                    selectMessage(message)
-                                } label: {
-                                    Image(
-                                        systemName:
-                                            selectedForDeletion.contains(
-                                                message
-                                            ) ? "trash.circle" : "circle"
-                                    )
-                                }
-                            }
-                            VStack {
-                                if !message.fromUser {
-                                    HStack(alignment: .center) {
-                                        if let character = message.character,
-                                            let avatar = character.avatar,
-                                            let image = UIImage(data: avatar)
-                                        {
-                                            Image(uiImage: image)
-                                                .resizable()
-                                                .scaledToFill()
-                                                .frame(
-                                                    width: 40,
-                                                    height: 40,
-                                                    alignment: .leading
-                                                )
-                                                .clipShape(Circle())
-                                                .glassEffect()
-                                        } else {
-                                            Image(
-                                                systemName: "person.circle.fill"
-                                            )
-                                            .resizable()
-                                            .scaledToFit()
-                                            .frame(
-                                                width: 40,
-                                                height: 40,
-                                                alignment: .leading
-                                            )
-                                            .clipShape(Circle())
-                                            .glassEffect()
-                                        }
-                                        Text(
-                                            message.character?.name
-                                                ?? "Unknown Character"
-                                        )
-                                        .font(.footnote)
-                                        .frame(
-                                            maxWidth: .infinity,
-                                            minHeight: 30,
-                                            alignment: .leading
-                                        )
-                                    }
-                                }
-                                MessageCell(
-                                    contentMessage: (message == messages.last
-                                        && !message.unwrappedContentAlternates
-                                            .isEmpty
-                                        && currentAlternateIndex >= 0
-                                        && currentAlternateIndex
-                                            < message.unwrappedContentAlternates
-                                            .count
-                                        ? message.unwrappedContentAlternates[
-                                            currentAlternateIndex
-                                        ].string : message.content)
-                                        .swapPlaceholders(
-                                            userName: chat.userName,
-                                            charName: message.character?.name,
-                                            userSettings: userSettings
-                                        ),
-                                    isCurrentUser: message.fromUser
-                                )
-                                .contextMenu {
-                                    Button(role: .destructive) {
-                                        if messages.last == message && !message.fromUser && currentAlternateIndex != -1 {
-                                            delete(contentAlternate: message.unwrappedContentAlternates[
-                                                currentAlternateIndex
-                                            ])
-                                        } else {
-                                            deleteMessage(message: message)
-                                        }
-                                    } label: {
-                                        Label(
-                                            (currentAlternateIndex != -1 ? "Delete Alternate" : "Delete Message"),
-                                            systemImage: "trash"
-                                        )
-                                    }
-
-                                    Button {
-                                        copyMessageText(message: message)
-                                    } label: {
-                                        Label(
-                                            "Copy Text",
-                                            systemImage: "doc.on.doc"
-                                        )
-                                    }
-
-                                    Button {
-                                        if messages.last == message && !message.fromUser && currentAlternateIndex != -1 {
-                                            alternateTextToEdit = message.unwrappedContentAlternates[
-                                                currentAlternateIndex
-                                            ].string
-                                            alternateBeingEdited = message.unwrappedContentAlternates[
-                                                currentAlternateIndex
-                                            ]
-                                            showAlternateTextEditor.toggle()
-                                        } else {
-                                            textToEdit = message.content
-                                            messageBeingEdited = message
-                                            showTextEditor.toggle()
-                                        }
-                                    } label: {
-                                        Label(
-                                            "Edit Text",
-                                            systemImage: "square.and.pencil"
-                                        )
-                                    }
-
-                                    Group {
-                                        if !message.fromUser,
-                                            message.request != nil
-                                        {
-                                            Button {
-                                                if messages.last == message && !message.fromUser && currentAlternateIndex != -1 {
-                                                    showRequestDetails(
-                                                        message.unwrappedContentAlternates[
-                                                            currentAlternateIndex
-                                                        ].request,
-                                                        message.unwrappedContentAlternates[
-                                                            currentAlternateIndex
-                                                        ].response
-                                                    )
-                                                } else {
-                                                    showRequestDetails(
-                                                        message.request,
-                                                        message.response
-                                                    )
-                                                }
-                                            } label: {
-                                                Label(
-                                                    "Generation Details",
-                                                    systemImage: "info.circle"
-                                                )
-                                            }
-                                        }
-                                    }
-                                }
-                                .padding(
-                                    message.fromUser ? .leading : .trailing,
-                                    message.fromUser ? 30 : 0
-                                )
-                            }
-                        }.id(message)
-                    }
-                    .padding([.leading, .trailing])
-                }
-                .onAppear {
-                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) {
-                        proxy.scrollTo(messages.last, anchor: .bottom)
-                    }
-                }
-                .onChange(of: messages.count) { oldValue, newValue in
-                    if (newValue - oldValue) > 0 {
-                        proxy.scrollTo(messages.last, anchor: .bottom)
-                    }
-                }
-            }
-            .defaultScrollAnchor(.bottom)
-            .navigationTitle(chat.name)
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbar {
-                if batchEditModeEnabled {
-                    ToolbarItemGroup(placement: .primaryAction) {
-                        Button("Delete Selected", systemImage: "trash") {
-                            batchDeleteMessages()
-                        }
-                        Button("Done") {
-                            selectedForDeletion.removeAll()
-                            batchEditModeEnabled = false
-                        }
-                    }
-                } else {
-                    ToolbarItemGroup(placement: .secondaryAction) {
-                        Button("Chat Settings", systemImage: "gearshape") {
-                            showingSettingsSheet.toggle()
-                        }
-                        Button("Chatlog View", systemImage: "list.clipboard") {
-                            showingChatlog.toggle()
-                        }
-                        if !showPendingMessage {
-                            Button("Batch Delete Mode", systemImage: "trash") {
-                                batchEditModeEnabled = true
-                            }
-                        }
-                        Menu {
-                            ForEach(chat.unwrappedCharacters, id: \.self) {
-                                character in
-                                Button(
-                                    "Edit \(character.name)",
-                                    systemImage: "person"
-                                ) {
-                                    selectedCharacter = character
-                                }
-                            }
-                            if chat.userCharacter != nil {
-                                Button(
-                                    "Edit \(chat.userCharacter!.name)",
-                                    systemImage: "person"
-                                ) {
-                                    selectedCharacter = chat.userCharacter!
-                                }
-                            } else if chat.userName == nil,
-                                userSettings.userCharacter != nil
-                            {
-                                Button(
-                                    "Edit \(userSettings.userCharacter!.name)",
-                                    systemImage: "person"
-                                ) {
-                                    selectedCharacter = userSettings
-                                        .userCharacter!
-                                }
-                            }
-                        } label: {
-                            Label("Characters", systemImage: "person.2")
-                        }
-                    }
-                }
-            }
-            .safeAreaBar(edge: .bottom) {
-                GlassEffectContainer {
-                    HStack(alignment: .bottom) {
-                        if isTextFieldFocused {
-                            Button {
-                                if newMessage == "" {
-                                    isTextFieldFocused.toggle()
-                                } else {
-                                    showingConfirmationDialog = true
-                                }
-                            } label: {
-                                Label("Close", systemImage: "xmark").labelStyle(
-                                    .iconOnly
-                                ).padding([.top, .bottom], 7)
-                            }
-                            .buttonStyle(.glass)
-                            .confirmationDialog(
-                                "Delete drafted message?",
-                                isPresented: $showingConfirmationDialog,
-                                titleVisibility: .visible
-                            ) {
-                                Button("Delete", role: .destructive) {
-                                    newMessage = ""
-                                    isTextFieldFocused.toggle()
-                                }
-                                Button("Cancel", role: .cancel) {
-                                    // do nothing
-                                    print("Canceled")
-                                }
-                            }
-                        }
-                        TextField(
-                            "AI Horde",
-                            text: $newMessage,
-                            axis: .vertical
-                        )
-                        .keyboardType(.asciiCapable)
-                        .lineLimit(5)
-                        .padding(
-                            EdgeInsets(
-                                top: 11,
-                                leading: 10,
-                                bottom: 11,
-                                trailing: 10
-                            )
-                        )
-                        .onSubmit {
-                            if newMessage != "" {
-                                requestMessage()
-                            }
-                        }
-                        .onReceive(keyboardPublisher) { value in
-                            if value {
-                                Log.debug("Keyboard Shown")
-                                keyboardShowing = true
-                            } else {
-                                Log.debug("Keyboard Hidden")
-                                keyboardShowing = false
-                            }
-                        }
-                        .focused($isTextFieldFocused)
-                        .glassEffect(
-                            .regular.interactive(),
-                            in: RoundedRectangle(cornerRadius: 20)
-                        )
-                        .glassEffectUnion(id: "1", namespace: unionNamespace)
-
-                        if !isTextFieldFocused, let lastMessage = messages.last,
-                            !lastMessage.fromUser
-                        {
-                            GlassEffectContainer {
-                                HStack {
-                                    if !lastMessage.unwrappedContentAlternates
-                                        .isEmpty
-                                    {
-                                        Button {
-                                            if !lastMessage
-                                                .unwrappedContentAlternates
-                                                .isEmpty
-                                            {
-                                                if currentAlternateIndex == -1 {
-                                                    currentAlternateIndex =
-                                                        lastMessage
-                                                        .unwrappedContentAlternates
-                                                        .count - 1
-                                                } else {
-                                                    currentAlternateIndex =
-                                                        currentAlternateIndex
-                                                        - 1
-                                                }
-                                            }
-                                        } label: {
-                                            Label(
-                                                "Undo",
-                                                systemImage:
-                                                    "arrowshape.backward"
-                                            ).labelStyle(.iconOnly).padding(
-                                                [.top, .bottom],
-                                                5
-                                            )
-                                        }
-                                        .buttonStyle(.glass)
-                                        .glassEffectUnion(
-                                            id: "back-retry",
-                                            namespace: buttonUnionNamespace
-                                        )
-                                        .disabled(
-                                            isPendingAlternate
-                                                || showPendingMessage
-                                        )
-                                    }
-
-                                    if !(messages.first == messages.last) {
-                                        Button {
-                                            getNewAlternateResponseToChat()
-                                        } label: {
-                                            Label("Reroll", systemImage: "dice")
-                                                .labelStyle(.iconOnly).padding(
-                                                    [.top, .bottom],
-                                                    5
-                                                )
-                                        }
-                                        .buttonStyle(.glass)
-                                        .glassEffectUnion(
-                                            id: "back-retry",
-                                            namespace: buttonUnionNamespace
-                                        )
-                                        .disabled(
-                                            isPendingAlternate
-                                                || showPendingMessage
-                                        )
-                                    }
-                                }
-                            }
-                        }
-
-                        if isTextFieldFocused {
-                            if newMessage != "" {
-                                Button {
-                                    requestMessage()
-                                } label: {
-                                    Label("Send", systemImage: "arrow.up")
-                                        .labelStyle(.iconOnly).padding(
-                                            [.top, .bottom],
-                                            5
-                                        )
-                                }
-                                .buttonStyle(.glassProminent)
-                                .disabled(
-                                    isPendingAlternate || showPendingMessage
-                                )
-                            }
-                        } else {
-                            Menu {
-                                Button {
-                                    requestMessage(imitation: true)
-                                } label: {
-                                    Label(
-                                        "Impersonate",
-                                        systemImage: "plus.message"
-                                    )
-                                }
-                                ForEach(chat.unwrappedCharacters, id: \.self) {
-                                    character in
-                                    Button {
-                                        requestMessage(fromCharacter: character)
-                                    } label: {
-                                        Label(
-                                            "\(character.name)",
-                                            systemImage: "plus.message"
-                                        )
-                                    }
-                                }
-                            } label: {
-                                Image(systemName: "plus.message").padding(
-                                    [.top, .bottom],
-                                    5
-                                )
-                            }
-                            .buttonStyle(.glass)
-                            .glassEffectUnion(
-                                id: "back-retry",
-                                namespace: buttonUnionNamespace
-                            )
-                            .disabled(isPendingAlternate)
-                            .disabled(showPendingMessage)
-                        }
-                    }.padding(
-                        (isTextFieldFocused && keyboardShowing
-                            ? [.leading, .trailing, .top, .bottom]
-                            : [.leading, .trailing, .top])
-                    )
-                }
-
-            }
-            .safeAreaBar(edge: .top) {
-                if showPendingMessage {
-                    Text(statusMessage)
-                        .padding()
-                        .foregroundStyle(.primary)
-                        .lineLimit(1)
-                        .minimumScaleFactor(0.8)
-                        .glassEffect()
-                        .padding([.leading, .trailing])
-                        .font(.footnote)
-                }
-            }
-            .sheet(isPresented: $showingSettingsSheet) {
-                ChatSettingsView(
-                    userSettings: userSettings,
-                    chat: chat,
-                    hordeRequest: viewModel.baseHordeRequest,
-                    hordeParams: viewModel.baseHordeParams
-                ).interactiveDismissDisabled()
-            }
-            .sheet(
-                isPresented: $showTextEditor,
-                content: {
-                    TextEditorView(text: $textToEdit)
-                }
-            )
-            .sheet(
-                isPresented: $showAlternateTextEditor,
-                content: {
-                    TextEditorView(text: $alternateTextToEdit)
-                }
-            )
-            .sheet(
-                isPresented: $showRequestDetails,
-                content: {
-                    GenerationDetailsView(
-                        responseDetails: $responseDetails,
-                        requestDetails: $requestDetails
-                    )
-                }
-            )
-            .sheet(
-                item: $selectedCharacter,
-                content: { character in
-                    CharacterView(character: character)
-                        .interactiveDismissDisabled()
-                }
-            )
-            .sheet(
-                isPresented: $showingChatlog,
-                content: {
-                    SelectableChatLogView(
-                        chat: chat,
-                        userSettings: userSettings
-                    )
-                }
-            )
-            .onChange(of: isTextFieldFocused) { _, newValue in
-                if newValue {
-                    withAnimation {
-                        proxy.scrollTo(messages.last, anchor: .bottom)
-                    }
-                }
-            }
-            .onChange(of: currentAlternateIndex) { _, _ in
-                withAnimation {
-                    proxy.scrollTo(messages.last, anchor: .bottom)
-                }
-            }
-            .onChange(of: showTextEditor) { _, newValue in
-                if !newValue {
-                    messageBeingEdited?.content = textToEdit
-                    try? modelContext.save()
-                }
-            }
-            .onChange(of: showAlternateTextEditor) { _, newValue in
-                if !newValue {
-                    alternateBeingEdited?.string = alternateTextToEdit
-                    try? modelContext.save()
-                }
-            }
-        }
-    }
-
-    func copyMessageText(message: ChatMessage) {
-        let pasteboard = UIPasteboard.general
-        pasteboard.string = message.content.swapPlaceholders(
-            userName: chat.userName,
-            charName: message.character?.name,
-            userSettings: userSettings
-        )
-    }
-
-    func batchDeleteMessages() {
-        for message in selectedForDeletion {
-            deleteMessage(message: message)
-        }
-        selectedForDeletion.removeAll()
-        batchEditModeEnabled = false
-    }
-
-    func deleteMessage(message: ChatMessage) {
-        if !message.unwrappedContentAlternates.isEmpty {
-            for alternate in message.unwrappedContentAlternates {
-                modelContext.delete(alternate)
-            }
-        }
-        modelContext.delete(message)
-    }
-    
-    func delete(contentAlternate: ContentAlternate) {
-        currentAlternateIndex -= 1
-        modelContext.delete(contentAlternate)
-    }
-
-    func selectMessage(_ message: ChatMessage) {
-        if selectedForDeletion.contains(message) {
-            selectedForDeletion.remove(message)
-        } else {
-            selectedForDeletion.insert(message)
-        }
-    }
-
-    func showRequestDetails(_ request: String?, _ response: String?) {
-        if let request, let response {
-            let jsonRequestData = Data(request.utf8)
-            requestDetails = jsonRequestData.printJson() ?? ""
-            let jsonResponseData = Data(response.utf8)
-            responseDetails = jsonResponseData.printJson() ?? ""
-            showRequestDetails.toggle()
-        }
-    }
-
-    func getNewAlternateResponseToChat() {
-        batchEditModeEnabled = false
-        if !showPendingMessage, let message = messages.last {
-            showPendingMessage.toggle()
-            isPendingAlternate.toggle()
-            statusMessage = "Requesting a new message..."
-            Task {
-                let response = await viewModel.getNewResponseToChat(
-                    statusMessage: $statusMessage,
-                    contentAlternate: true
-                )
-                let newAlternate = ContentAlternate(
-                    string: response.text,
-                    request: response.request,
-                    response: response.response
-                )
-                message.contentAlternates?.append(newAlternate)
-                chat.dateUpdated = .now
-                DispatchQueue.main.asyncAfter(deadline: .now() + 1) {
-                    currentAlternateIndex =
-                        message.unwrappedContentAlternates.count - 1
-                    showPendingMessage.toggle()
-                    isPendingAlternate.toggle()
-                }
-            }
-        }
-    }
-
-    func requestMessage(
-        fromCharacter: Character? = nil,
-        imitation: Bool = false
-    ) {
-        batchEditModeEnabled = false
-        if currentAlternateIndex >= 0, let currentMessage = messages.last,
-            currentMessage.unwrappedContentAlternates.count
-                >= currentAlternateIndex + 1, !currentMessage.fromUser
-        {
-            let alternateContent = currentMessage.unwrappedContentAlternates[
-                currentAlternateIndex
-            ]
-            let originalContent = currentMessage.content
-            let originalRequest = currentMessage.request
-            let originalResponse = currentMessage.response
-
-            currentMessage.content = alternateContent.string
-            currentMessage.request = alternateContent.request
-            currentMessage.response = alternateContent.response
-            alternateContent.string = originalContent
-            alternateContent.response = originalResponse
-            alternateContent.request = originalRequest
-            currentAlternateIndex = -1
-        }
-        if !newMessage.isEmpty {
-            let newUserMessage = ChatMessage(
-                content: newMessage,
-                fromUser: true,
-                chat: chat
-            )
-            modelContext.insert(newUserMessage)
-            chat.dateUpdated = .now
-            statusMessage = "Sending message..."
-        } else {
-            statusMessage = "Requesting a new message..."
-        }
-        showPendingMessage.toggle()
-        newMessage = ""
-        currentAlternateIndex = -1
-        isTextFieldFocused = false
-        try? modelContext.save()
-        Task {
-            let response = await viewModel.getNewResponseToChat(
-                statusMessage: $statusMessage,
-                character: fromCharacter,
-                imitation: imitation
-            )
-            let newResponseMessage = ChatMessage(
-                content: response.text,
-                fromUser: imitation,
-                chat: chat,
-                character: imitation ? nil : response.character,
-                request: response.request,
-                response: response.response
-            )
-            chat.dateUpdated = Date.now
-            modelContext.insert(newResponseMessage)
-            chat.dateUpdated = .now
-            try? modelContext.save()
-            showPendingMessage.toggle()
-        }
-    }
-}
-
-#Preview {
-    let config = ModelConfiguration(isStoredInMemoryOnly: true)
-    let container = try! ModelContainer(for: Chat.self, configurations: config)
-    let modelContext = container.mainContext
-    let character = Character(
-        name: "Bradley Root",
-        characterDescription: "Bradley is a software engineer",
-        personality: "",
-        firstMessage: "Hi! I'm Bradley!",
-        exampleMessage: "",
-        scenario: "",
-        creatorNotes: "",
-        systemPrompt: "",
-        postHistoryInstructions: "",
-        alternateGreetings: [],
-        tags: [],
-        creator: "Brad Root",
-        characterVersion: "main",
-        chubId: "",
-        avatar: UIImage(named: "brad-drawn")!.pngData()!
-    )
-    container.mainContext.insert(character)
-    let chat = Chat(name: "Chat Name", characters: [character])
-    container.mainContext.insert(chat)
-    for i in 1..<10 {
-        let message = ChatMessage(
-            content:
-                "Lorem ipsum dolor sit amet. {{user}}? {{char}}? {{User}}? {{Char}}? consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-            fromUser: i % 2 == 0 ? true : false,
-            chat: chat,
-            character: character
-        )
-        message.contentAlternates = [
-            ContentAlternate(
-                string: "Blah",
-                message: message,
-                request: nil,
-                response: nil
-            ),
-            ContentAlternate(
-                string: "Blah 2",
-                message: message,
-                request: nil,
-                response: nil
-            ),
-            ContentAlternate(
-                string: "Blah 3",
-                message: message,
-                request: nil,
-                response: nil
-            ),
-        ]
-        container.mainContext.insert(message)
-    }
-    return NavigationStack {
-        NewChatView(
-            for: chat,
-            modelContext: modelContext,
-            userSettings: UserSettings(
-                userCharacter: nil,
-                defaultUserName: "Seymour"
-            )
-        )
-        .modelContainer(container)
-        .navigationTitle("Seymour")
-    }
-}
